added pidalt stuff into cli added a modifier for set (type "set *") to see min/max values for each parameter. this is for gui stuff bumped config VERSION fixed 0xffff - now stuff in drv_pwm, after it was mentioned to me for the  3rd time :p replaced some uint8_t -> uint32_t for loop counters - saving 4 to 8 bytes of flash each time. thanks goes to thee 35+ years of C experience kicad guy. turned bitfiends into a regular struct - instant 100byte flash size reduction
git-svn-id: https://afrodevices.googlecode.com/svn/trunk/baseflight@174 7c89a4a9-59b9-e629-4cfe-3a2d53b20e61

+++ src/imu.c
@@ -34,7 +34,7 @@ void imuInit(void)
 
 void computeIMU(void)
 {
-    uint8_t axis;
+    uint32_t axis;
     static int16_t gyroADCprevious[3] = { 0, 0, 0 };
     int16_t gyroADCp[3];
     int16_t gyroADCinter[3];
@@ -180,7 +180,7 @@ static int16_t _atan2f(float y, float x)
 
 static void getEstimatedAttitude(void)
 {
-    uint8_t axis;
+    uint32_t axis;
     int32_t accMag = 0;
     static t_fp_vector EstG;
     static t_fp_vector EstM;
@@ -265,10 +265,10 @@ static void getEstimatedAttitude(void)
 
 void getEstimatedAltitude(void)
 {
-    uint8_t index;
+    uint32_t index;
     static uint32_t deadLine = INIT_DELAY;
     static int16_t BaroHistTab[BARO_TAB_SIZE];
-    static int8_t BaroHistIdx;
+    static uint32_t BaroHistIdx;
     static int32_t BaroHigh = 0;
     static int32_t BaroLow = 0;
     int32_t temp32;
